Add DropShadow.color: [f32;4]
