+++ src/composition.rs
@@ -2,6 +2,7 @@ use crate::error::MakuError;
 use crate::io;
 use crate::programs;
 use crate::target;
+use crate::value;
 
 /// Represents different types of filters that can be applied to an image
 pub enum Filter {
@@ -18,50 +19,8 @@ pub enum Filter {
     /// A shader filter, containing a program
     Shader {
         program: three_d::Program,
-        uniforms: Vec<(String, UniformValue)>,
+        uniforms: Vec<(String, value::UniformValue)>,
     },
-}
-
-pub enum UniformValue {
-    Float(f32),
-    Vector2(three_d::Vector2<f32>),
-    Vector3(three_d::Vector3<f32>),
-    Vector4(three_d::Vector4<f32>),
-}
-
-impl UniformValue {
-    pub fn apply(&self, program: &three_d::Program, name: &str) {
-        match self {
-            UniformValue::Float(v) => program.use_uniform(name, v),
-            UniformValue::Vector2(v) => program.use_uniform(name, v),
-            UniformValue::Vector3(v) => program.use_uniform(name, v),
-            UniformValue::Vector4(v) => program.use_uniform(name, v),
-        };
-    }
-}
-
-impl From<f32> for UniformValue {
-    fn from(value: f32) -> Self {
-        Self::Float(value)
-    }
-}
-
-impl From<(f32, f32)> for UniformValue {
-    fn from(value: (f32, f32)) -> Self {
-        Self::Vector2(three_d::Vector2::new(value.0, value.1))
-    }
-}
-
-impl From<(f32, f32, f32)> for UniformValue {
-    fn from(value: (f32, f32, f32)) -> Self {
-        Self::Vector3(three_d::Vector3::new(value.0, value.1, value.2))
-    }
-}
-
-impl From<(f32, f32, f32, f32)> for UniformValue {
-    fn from(value: (f32, f32, f32, f32)) -> Self {
-        Self::Vector4(three_d::Vector4::new(value.0, value.1, value.2, value.3))
-    }
 }
 
 // Composition
@@ -311,12 +270,20 @@ fn load_shader_filter(
             include_str!("./presets/gaussian_blur.frag").to_string(),
             vec![("u_radius".to_string(), (*radius).into())],
         ),
-        io::IoFilter::DropShadow { radius, offset } => (
+        io::IoFilter::DropShadow {
+            radius,
+            offset,
+            color,
+        } => (
             include_str!("./presets/drop_shadow.vert").to_string(),
             include_str!("./presets/drop_shadow.frag").to_string(),
             vec![
                 ("u_radius".to_string(), (*radius).into()),
                 ("u_offset".to_string(), (offset[0], offset[1]).into()),
+                (
+                    "u_color".to_string(),
+                    (color[0], color[1], color[2], color[3]).into(),
+                ),
             ],
         ),
         io::IoFilter::Composition(..) | io::IoFilter::Image { .. } => unreachable!(),
